moved trending to own component and added to top level so is on all pages
