+++ react-app/src/index.css
@@ -123,3 +123,67 @@ button:hover{
   position: relative;
   left: 4%;
 }
+
+#all-news-container{
+  position: absolute;
+  display: inline-flex;
+  flex-direction: column;
+  width: 20%;
+  overflow-wrap: normal;
+  top: 30px;
+  right:12%;
+  background-color: #f1f4f5;
+  border-radius: 12px;
+  padding: 5px;
+}
+
+#each-news-container{
+  display: inline;
+  height: 80px;
+  width: 90%;
+}
+
+#each-news-container:hover{
+  background-color: #eeeeee;
+}
+
+#news-headline{
+  margin: 0;
+}
+
+#news-image-provider{
+  margin: 0;
+  font-size: x-small;
+}
+
+.news-buttons{
+  background-color: #f1f4f5;
+  height: 40px;
+  text-align: left;
+  border: 0;
+  color: #269ff0;
+  width: 100%;
+  border-bottom-right-radius: 12px;
+  border-bottom-left-radius: 12px;
+}
+
+.news-buttons:hover{
+  background-color: #eeeeee;
+  cursor: pointer;
+}
+
+#news-image{
+  height: 60px;
+  width: 60px;
+  border-radius: 12px;
+}
+
+#news-image-info{
+  text-align: end;
+  width: 100%;
+  margin: 0;
+}
+
+#news-text{
+  width: 100%;
+}
